Métodos de la clase Persona implementados, creación de constructores en la clase Factura y correcciones en algunas partes de la estructura lógica
Se implementan los métodos correspondientes a la clase persona, también se implementa el método mostrarFactura en la clase Transportadora, se crean los constructores de la clase Factura y se corrigen algunas partes de la estructura lógica del código

+++ src/gestorAplicacion/administrativo/Viaje.java
@@ -143,7 +143,7 @@ public class Viaje {
      * Establece o modifica la tarifa del viaje.
      * @param tarifa La tarifa del viaje.
      */
-    public void setTarifa(int tarifa) {
+    public void setTarifa(double tarifa) {
         this.tarifa = tarifa;
     }
